refactor(router): move rules into their own folder
